[CHORE](foundation-api): Share whoami_and_authorize helper
Move the two-step auth dance out of `init.rs` and into a new
`routes/whoami.rs` so future handlers in the foundation route surface
can reuse it without re-inlining the gotcha (Cloud authz 403s when
`resource.tenant != identity.tenant`, including `tenant: None`).

The helper now returns the full `GetUserIdentityResponse` rather than
just the tenant, so callers that also need `user_id` don't have to
make a second `get_user_identity` call.

`init.rs` keeps its existing behavior; this is a pure refactor.
