+++ rust/foundation-api/src/routes/init.rs
@@ -13,11 +13,9 @@ use serde::Serialize;
 use std::collections::HashMap;
 use uuid::Uuid;
 
+use super::whoami::whoami_and_authorize;
 use crate::{
-    auth::{AuthError, AuthenticateAndAuthorize, AuthzAction, AuthzResource},
-    config::FoundationConfig,
-    errors::ServerError,
-    server::FoundationApiServer,
+    auth::AuthzAction, config::FoundationConfig, errors::ServerError, server::FoundationApiServer,
 };
 
 #[derive(Serialize)]
@@ -41,7 +39,9 @@ pub async fn foundation_init(
     headers: HeaderMap,
     State(server): State<FoundationApiServer>,
 ) -> Result<Json<FoundationInitResponse>, ServerError> {
-    let tenant = whoami_and_authorize(&*server.auth, &headers, AuthzAction::InitFoundation).await?;
+    let tenant = whoami_and_authorize(&*server.auth, &headers, AuthzAction::InitFoundation)
+        .await?
+        .tenant;
 
     let _guard =
         server.scorecard_request(&["op:foundation_init", &format!("tenant:{}", tenant)])?;
@@ -276,106 +276,10 @@ async fn ensure_collection(
     Ok(collection)
 }
 
-/// Resolve the caller's tenant from the auth token, then check that the
-/// caller is allowed to perform `action` against that tenant.
-///
-/// The two-step shape exists because the Cloud `authenticate_and_authorize`
-/// impl enforces `resource.tenant == user_identity.tenant` (returns 403 on
-/// mismatch, including `resource.tenant == None`). The Noop impl ignores
-/// resource entirely, which is why a single-call handler that passed
-/// `tenant: None` looked fine in tests but 403'd under Cloud auth.
-async fn whoami_and_authorize(
-    auth: &dyn AuthenticateAndAuthorize,
-    headers: &HeaderMap,
-    action: AuthzAction,
-) -> Result<String, AuthError> {
-    let identity = auth.get_user_identity(headers).await?;
-    let tenant = identity.tenant.clone();
-    auth.authenticate_and_authorize(
-        headers,
-        action,
-        AuthzResource {
-            tenant: Some(tenant.clone()),
-            database: None,
-            collection: None,
-        },
-    )
-    .await?;
-    Ok(tenant)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use chroma_api_types::GetUserIdentityResponse;
     use chroma_types::SegmentType;
-    use std::collections::HashSet;
-    use std::future::{ready, Future};
-    use std::pin::Pin;
-    use std::sync::Mutex;
-
-    /// Fake `AuthenticateAndAuthorize` that returns a fixed tenant from
-    /// `get_user_identity` and records the `AuthzResource` passed to
-    /// `authenticate_and_authorize` so tests can assert on it.
-    struct FakeAuth {
-        tenant: String,
-        captured_action: Mutex<Option<AuthzAction>>,
-        captured_resource: Mutex<Option<AuthzResource>>,
-    }
-
-    impl FakeAuth {
-        fn new(tenant: &str) -> Self {
-            Self {
-                tenant: tenant.to_string(),
-                captured_action: Mutex::new(None),
-                captured_resource: Mutex::new(None),
-            }
-        }
-
-        fn identity(&self) -> GetUserIdentityResponse {
-            GetUserIdentityResponse {
-                user_id: String::new(),
-                tenant: self.tenant.clone(),
-                databases: HashSet::new(),
-            }
-        }
-    }
-
-    impl AuthenticateAndAuthorize for FakeAuth {
-        fn authenticate_and_authorize(
-            &self,
-            _headers: &axum::http::HeaderMap,
-            action: AuthzAction,
-            resource: AuthzResource,
-        ) -> Pin<Box<dyn Future<Output = Result<GetUserIdentityResponse, AuthError>> + Send>>
-        {
-            *self.captured_action.lock().unwrap() = Some(action);
-            *self.captured_resource.lock().unwrap() = Some(resource);
-            let identity = self.identity();
-            Box::pin(ready(Ok(identity)))
-        }
-
-        fn authenticate_and_authorize_collection(
-            &self,
-            _headers: &axum::http::HeaderMap,
-            _action: AuthzAction,
-            _resource: AuthzResource,
-            _collection: chroma_types::Collection,
-        ) -> Pin<Box<dyn Future<Output = Result<GetUserIdentityResponse, AuthError>> + Send>>
-        {
-            let identity = self.identity();
-            Box::pin(ready(Ok(identity)))
-        }
-
-        fn get_user_identity(
-            &self,
-            _headers: &axum::http::HeaderMap,
-        ) -> Pin<Box<dyn Future<Output = Result<GetUserIdentityResponse, AuthError>> + Send>>
-        {
-            let identity = self.identity();
-            Box::pin(ready(Ok(identity)))
-        }
-    }
 
     #[test]
     fn foundation_schema_has_sparse_vector_index() {
@@ -436,34 +340,5 @@ mod tests {
             "plan must include a SPANN vector segment, got: {:?}",
             plan.segments.iter().map(|s| &s.r#type).collect::<Vec<_>>()
         );
-    }
-
-    #[tokio::test]
-    async fn whoami_and_authorize_passes_resolved_tenant_to_authz() {
-        let fake = FakeAuth::new("team_abc");
-        let headers = HeaderMap::new();
-
-        let tenant = whoami_and_authorize(&fake, &headers, AuthzAction::InitFoundation)
-            .await
-            .expect("auth should succeed");
-
-        assert_eq!(tenant, "team_abc");
-        let captured_action = fake
-            .captured_action
-            .lock()
-            .unwrap()
-            .expect("authenticate_and_authorize should have been called");
-        assert_eq!(captured_action, AuthzAction::InitFoundation);
-        let captured = fake
-            .captured_resource
-            .lock()
-            .unwrap()
-            .clone()
-            .expect("authenticate_and_authorize should have been called");
-        // Regression: before this fix the handler passed `tenant: None`,
-        // which the Cloud authz impl always rejects with 403.
-        assert_eq!(captured.tenant, Some("team_abc".to_string()));
-        assert_eq!(captured.database, None);
-        assert_eq!(captured.collection, None);
     }
 }
